// Blogging Project Angular and asp.net/BlogProjecr/UI/BlogUi/src/app/models/blog-post.model.ts
import { Category } from "./Category-model";

export interface BlogPost {
    id: string;
    title: string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    author: string;
    publishedDate: Date;
    isVisible: boolean;
    categories: Category[];
  }